Add files via upload
This is the code/algorithm that I am messing around with to parallelize it. It has the memory allocation and everything like that that we need for the code. It is the algorithm from that one website we shared. There is no example of for making it parallel but basically I am trying to send to function sort and then within that function it calls the GPU kernel and returns the count in the buckets. The CPU reorganizes the buckets and creates the sorted array. 

*NOTE: the radixsort function currently is the CPU version of the algorithm. Running time function allows us to see how fast each of the functions are running. To run, also include the number of integers. For example: ./Radix_Sort 5